pan seg as 32bit

+++ empanada/inference/rle.py
@@ -103,7 +103,7 @@ def rle_seg_to_pan_seg(
 
     """
     # convert from dense panoptic seg to sparse rle segment class
-    pan_seg = np.zeros(shape, dtype='int').ravel()
+    pan_seg = np.zeros(shape, dtype=np.uint32).ravel()
 
     for instance_attrs in rle_seg.values():
         for object_id, attrs in instance_attrs.items():
